Refactor IO, change name conventions
